various bug fixes

+++ aevnmt/models/generative.py
@@ -312,7 +312,7 @@ class CorrelatedCategoricalsLM(GenerativeLM):
 
     def log_prob(self, likelihood: Categorical, x):
         # [B, Tx] -> [B]
-        return (likelihood.log_prob(x) * (x != self.pad_idx)).sum(-1)
+        return (likelihood.log_prob(x) * (x != self.pad_idx).float()).sum(-1)
 
     def sample(self, z, max_len=100, greedy=False, state=dict()):
         """
